wms-backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.11.10




